+++ axelor-account/src/main/java/com/axelor/apps/account/db/repo/InvoiceManagementRepository.java
@@ -57,6 +57,7 @@ public class InvoiceManagementRepository extends InvoiceRepository {
     copy.setPartnerAccount(null);
     copy.setJournal(null);
     copy.clearInvoicePaymentList();
+    copy.setPrintedPDF(null);
 
     return copy;
   }