+++ axelor-production/src/main/java/com/axelor/apps/production/service/costsheet/CostSheetLineServiceImpl.java
@@ -18,10 +18,14 @@
 package com.axelor.apps.production.service.costsheet;
 
 import com.axelor.apps.base.db.Company;
+import com.axelor.apps.base.db.Currency;
+import com.axelor.apps.base.db.Partner;
 import com.axelor.apps.base.db.Product;
 import com.axelor.apps.base.db.Unit;
 import com.axelor.apps.base.db.repo.ProductRepository;
 import com.axelor.apps.base.db.repo.UnitRepository;
+import com.axelor.apps.base.service.CurrencyService;
+import com.axelor.apps.base.service.ShippingCoefService;
 import com.axelor.apps.base.service.UnitConversionService;
 import com.axelor.apps.production.db.BillOfMaterial;
 import com.axelor.apps.production.db.CostSheetGroup;
@@ -56,6 +60,8 @@ public class CostSheetLineServiceImpl implements CostSheetLineService {
   protected UnitRepository unitRepo;
   protected WeightedAveragePriceService weightedAveragePriceService;
   protected UnitCostCalcLineServiceImpl unitCostCalcLineServiceImpl;
+  protected CurrencyService currencyService;
+  protected ShippingCoefService shippingCoefService;
 
   @Inject
   public CostSheetLineServiceImpl(
@@ -64,13 +70,17 @@ public class CostSheetLineServiceImpl implements CostSheetLineService {
       UnitConversionService unitConversionService,
       UnitRepository unitRepo,
       WeightedAveragePriceService weightedAveragePriceService,
-      UnitCostCalcLineServiceImpl unitCostCalcLineServiceImpl) {
+      UnitCostCalcLineServiceImpl unitCostCalcLineServiceImpl,
+      CurrencyService currencyService,
+      ShippingCoefService shippingCoefService) {
     this.appProductionService = appProductionService;
     this.costSheetGroupRepository = costSheetGroupRepository;
     this.unitConversionService = unitConversionService;
     this.unitRepo = unitRepo;
     this.weightedAveragePriceService = weightedAveragePriceService;
     this.unitCostCalcLineServiceImpl = unitCostCalcLineServiceImpl;
+    this.currencyService = currencyService;
+    this.shippingCoefService = shippingCoefService;
   }
 
   public CostSheetLine createCostSheetLine(
@@ -205,7 +215,8 @@ public class CostSheetLineServiceImpl implements CostSheetLineService {
             break;
           }
         }
-
+        // If we didn't have a computed price in cost calculation session, so we compute the price
+        // from its bill of material
       case CostSheetService.ORIGIN_BILL_OF_MATERIAL:
         costPrice =
             this.getComponentCostPrice(
@@ -216,15 +227,14 @@ public class CostSheetLineServiceImpl implements CostSheetLineService {
         costPrice = BigDecimal.ZERO;
     }
 
+    costPrice = costPrice.multiply(consumptionQty);
     costPrice =
-        unitConversionService
-            .convert(
-                unit,
-                product.getUnit(),
-                costPrice,
-                appProductionService.getNbDecimalDigitForUnitPrice(),
-                product)
-            .multiply(consumptionQty);
+        unitConversionService.convert(
+            unit,
+            product.getUnit(),
+            costPrice,
+            appProductionService.getNbDecimalDigitForUnitPrice(),
+            product);
 
     List<CostSheetLine> costSheetLineList =
         parentCostSheetLine.getCostSheetLineList() != null
@@ -260,29 +270,61 @@ public class CostSheetLineServiceImpl implements CostSheetLineService {
   }
 
   protected BigDecimal getComponentCostPrice(
-      Product product, int componentsValuationMethod, Company company) {
+      Product product, int componentsValuationMethod, Company company) throws AxelorException {
 
     BigDecimal price = null;
+    Currency companyCurrency = company.getCurrency();
 
     if (componentsValuationMethod == ProductRepository.COMPONENTS_VALUATION_METHOD_AVERAGE) {
       price = weightedAveragePriceService.computeAvgPriceForCompany(product, company);
+
+      if (price == null || price.compareTo(BigDecimal.ZERO) == 0) {
+        price = product.getCostPrice();
+      }
     } else if (componentsValuationMethod == ProductRepository.COMPONENTS_VALUATION_METHOD_COST) {
       price = product.getCostPrice();
+
+      if (price == null || price.compareTo(BigDecimal.ZERO) == 0) {
+        price = weightedAveragePriceService.computeAvgPriceForCompany(product, company);
+      }
     }
-    if (price == null || price.compareTo(BigDecimal.ZERO) == 0) {
-      price = product.getCostPrice();
-    }
-    if (price == null || price.compareTo(BigDecimal.ZERO) == 0) {
-      price = weightedAveragePriceService.computeAvgPriceForCompany(product, company);
-    }
+
     if (price == null || price.compareTo(BigDecimal.ZERO) == 0) {
       price = product.getPurchasePrice();
-    }
-    if (price == null || price.compareTo(BigDecimal.ZERO) == 0) {
-      for (SupplierCatalog supplierCatalog : product.getSupplierCatalogList()) {
-        if (BigDecimal.ZERO.compareTo(supplierCatalog.getPrice()) < 0) {
-          price = supplierCatalog.getPrice();
-          break;
+
+      BigDecimal shippingCoef =
+          shippingCoefService.getShippingCoef(
+              product, product.getDefaultSupplierPartner(), company, new BigDecimal(9999999));
+
+      price = product.getPurchasePrice().multiply(shippingCoef);
+
+      price =
+          currencyService.getAmountCurrencyConvertedAtDate(
+              product.getPurchaseCurrency(),
+              companyCurrency,
+              price,
+              appProductionService.getTodayDate());
+
+      if (price == null || price.compareTo(BigDecimal.ZERO) == 0) {
+        for (SupplierCatalog supplierCatalog : product.getSupplierCatalogList()) {
+          if (BigDecimal.ZERO.compareTo(supplierCatalog.getPrice()) < 0) {
+            price = supplierCatalog.getPrice();
+            Partner supplierPartner = supplierCatalog.getSupplierPartner();
+            if (supplierPartner != null) {
+              shippingCoef =
+                  shippingCoefService.getShippingCoef(
+                      product, supplierPartner, company, new BigDecimal(9999999));
+              price = price.multiply(shippingCoef);
+
+              price =
+                  currencyService.getAmountCurrencyConvertedAtDate(
+                      supplierPartner.getCurrency(),
+                      companyCurrency,
+                      price,
+                      appProductionService.getTodayDate());
+            }
+            break;
+          }
         }
       }
     }