+++ axelor-purchase/src/main/java/com/axelor/apps/purchase/service/PurchaseOrderLineTaxService.java
@@ -102,7 +102,7 @@ public class PurchaseOrderLineTaxService {
                 map.get(taxEquiv.getReverseChargeTax().getActiveTaxLine());
 
             purchaseOrderLineRC.setExTaxBase(
-                purchaseOrderLineRC.getExTaxBase().subtract(purchaseOrderLine.getExTaxTotal()));
+                purchaseOrderLineRC.getExTaxBase().add(purchaseOrderLine.getExTaxTotal()));
 
             purchaseOrderLineRC.setReverseCharged(true);