+++ axelor-purchase/src/main/java/com/axelor/apps/purchase/service/PurchaseOrderServiceImpl.java
@@ -271,7 +271,7 @@ public class PurchaseOrderServiceImpl implements PurchaseOrderService {
                 : "");
 
     ReportFactory.createReport(IReport.PURCHASE_ORDER, title + "-${date}")
-        .addParam("PurchaseOrderId", purchaseOrder.getId().toString())
+        .addParam("PurchaseOrderId", purchaseOrder.getId())
         .addParam("Locale", language)
         .toAttach(purchaseOrder)
         .generate()