+++ axelor-sale/src/main/java/com/axelor/apps/sale/service/saleorder/SaleOrderWorkflowServiceImpl.java
@@ -177,6 +177,7 @@ public class SaleOrderWorkflowServiceImpl implements SaleOrderWorkflowService {
   @Transactional(rollbackOn = {AxelorException.class, RuntimeException.class})
   public void completeSaleOrder(SaleOrder saleOrder) throws AxelorException {
     saleOrder.setStatusSelect(SaleOrderRepository.STATUS_ORDER_COMPLETED);
+    saleOrder.setOrderBeingEdited(false);
 
     saleOrderRepo.save(saleOrder);
   }