+++ CHANGELOG.md
@@ -77,6 +77,13 @@
 - DataBackup : update importId when its null with format IDDDMMYYHHMM
 - Project : Added dashlet view of contract invoices
 - Quality : Major updates in axelor-quality module
+- Business Project : Addition of new field statusSelect in InvoicingProject.
+- Generating supplier order from partner form.
+- Mobile App : Add configuration for quality app on mobile app
+- INVOICE : new process to print
+- QUALITY CONTROL : New report to print
+- PERMISSIONS : Display dashlets for groups/roles/users using the permission in permission and meta permission form views.
+- AppBase : Addition of new boolean 'enableMultiCompany' also show/hide company column in all grid-view base on it. 
 
 ## Improvements
 - Contract: added button to manually close contract if termination date was set in the future
@@ -103,6 +110,9 @@
 - Employment Contract : Add buttons to "Tools" in menubar and set form width to large
 - TeamsTask : New menu titles
 - Contract : partner / project filters improved
+- TeamTask : Set default typeSelect = 'task'
+- APP BUILDER : Remove JsonCreator object
+- TeamTask : copy object with empty fields
 
 ## Bug Fixes
 - TeamTask : Resolve NPE on save of new team task.
@@ -115,5 +125,8 @@
 - Accounting situation : added two fields (insurCreditDate and description) to manage the Credit Insurance.
 - Employee : Fix issue of not saving each phase of creation process
 - TeamTask : Fix issue of creating new record from calendar view
+- Marketing : Fix error when trying to generate an event per target
+- Contract : Fix import error in data-init. 
+- Sale Quotation : Move menu from CRM to Sale module.
 
 [Unreleased 5.2.0]: https://github.com/axelor/axelor-business-suite/compare/dev...wip